+++ Backend/config/config.json
@@ -1,5 +1,6 @@
 {
-  "PORT": 3000,
+  "PORT": 5000,
   "URL": "mongodb://localhost/E_ulavan",
   "JWTSECREATEKEY": "reshma_rohith"
 }
+
