+++ frontend/src/Router.js
@@ -1,12 +1,9 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Registration from "./components/Registration";
-
 import HomeCard from "./components/HomeCard";
-import Home from './components/Home'
+import Home from "./components/Home";
 
-
-import Home from "./components/home";
 
 const Routering = () => {
   return (
@@ -14,7 +11,7 @@ const Routering = () => {
     <Router>
       <Routes>
 
-        <Route exact path="/" element={<Home />} />
+        {/* <Route exact path="/" element={<Home />} /> */}
 
         <Route exact path="/" element={<Registration />} />
         <Route path="/home" element={<Home/>} />
