+++ frontend/src/components/HomeCard.js
@@ -1,38 +1,82 @@
-import React from "react";
-import { Grid, Button } from "@mui/material";
-import {
-  Card,
-  CardContent,
-  CardActionArea,
-  Typography,
-} from "@mui/material";
+import React, { useState } from "react";
+import { Grid , Button} from "@mui/material";
+import { Card, CardContent, CardActionArea, Typography } from "@mui/material";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
+
+import { Modal, Form } from "react-bootstrap";
+
+import axios from "axios";
+
 const HomeCard = () => {
+  //----------this is for manufacturer details--------------//
+  const [show, setShow] = useState(false);
+
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
+
+  const [name, setName] = useState("");
+  const [address, setAddress] = useState("");
+  const [mobileNumber, setMobileNumber] = useState("");
+  const [bio, setBio] = useState("");
+  //  for  all fields
+  const [alert, setAlert] = useState(false);
+// ====================================================//
+
+//========this alerterror for fill all the fields=========//
+  const alerterror = () => {
+    if (name !== "" && address !== "" && mobileNumber !== "" && bio !== "") {
+      const former = {
+        name: name,
+        address: address,
+        mobileNumber: mobileNumber,
+        bio: bio,
+      };
+// ==========this is for  api============//
+      const url = "http://localhost:5000/post";
+
+      axios.post(url, former).then((res) => {
+        console.log(res);
+      });
+
+      console.log(former);
+      setAlert(false);
+    } else {
+      setAlert(true);
+    }
+  };
   return (
-    <Grid
-      container
-      style={{
-        marginTop: "100px",
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Grid item md={6}>
-        <Card style={{ backgroundColor: "#1DD455" }}>
-          <CardActionArea>
-            <CardContent>
-              <Typography
-                variant="h5"
-                component={"h5"}
-                style={{ textAlign: "center", color: "#FFFFFF" }}
-              >
-                CHOOSE YOUR ROLE
-              </Typography>
+    <>
+    {/* this card for coose role */}
+      <Grid
+        container
+        style={{
+          marginTop: "100px",
+          justifyContent: "center",
+          alignItems: "center",
+        }}
+      >
+        <Grid item md={6}>
+          <Card style={{ backgroundColor: "#1DD455" }}>
+            <CardActionArea>
+              <CardContent>
+                <Typography
+                  variant="h5"
+                  component={"h5"}
+                  style={{ textAlign: "center", color: "#FFFFFF" }}
+                >
+                  CHOOSE YOUR ROLE
+                </Typography>
                 <Button
+                  onClick={handleShow}
                   variant="contained"
                   color="secondary"
                   startIcon={<AccountCircleIcon />}
-                  sx={{ display: "inline-block", width: "70%",marginLeft:'80px',marginTop:'40px' }}
+                  sx={{
+                    display: "inline-block",
+                    width: "70%",
+                    marginLeft: "80px",
+                    marginTop: "40px",
+                  }}
                 >
                   MANUFACTURER
                 </Button>
@@ -40,15 +84,83 @@ const HomeCard = () => {
                   variant="contained"
                   color="secondary"
                   startIcon={<AccountCircleIcon />}
-                  sx={{ display: "inline-block", width: "70%",marginLeft:'80px',marginTop:'40px',marginBottom:'40px' }}
+                  sx={{
+                    display: "inline-block",
+                    width: "70%",
+                    marginLeft: "80px",
+                    marginTop: "40px",
+                    marginBottom: "40px",
+                  }}
                 >
                   BUYER
                 </Button>
-            </CardContent>
-          </CardActionArea>
-        </Card>
+              </CardContent>
+            </CardActionArea>
+          </Card>
+        </Grid>
       </Grid>
-    </Grid>
+{/* this is manufacturer details */}
+      <Modal
+        show={show}
+        onHide={handleClose}
+        backdrop="static"
+        keyboard={false}
+      >
+        <Modal.Header closeButton>
+          <Modal.Title>ENTER YOUR DETAILS</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <Form>
+            <h6 style={{ displayFlex: "inline-flex" }}>
+              name :
+              <Form.Control
+                onChange={(e) => setName(e.target.value)}
+                placeholder="enter your name...."
+              ></Form.Control>
+            </h6>
+
+            <h6>
+              ADDRESS :{" "}
+              <Form.Control
+                onChange={(e) => setAddress(e.target.value)}
+                style={{ padding: "30px" }}
+                placeholder="enter your address here...."
+              ></Form.Control>{" "}
+            </h6>
+
+            <h6>
+              MOBILE NUMBER :
+              <Form.Control
+                onChange={(e) => setMobileNumber(e.target.value)}
+                placeholder="9876543210"
+              ></Form.Control>{" "}
+            </h6>
+
+            <h6>
+              BIO :{" "}
+              <Form.Control
+                onChange={(e) => setBio(e.target.value)}
+                style={{ padding: "30px" }}
+                placeholder="enter your bio here...."
+              ></Form.Control>{" "}
+            </h6>
+          </Form>
+          {alert === true && (
+            <p style={{ color: "red" }}>
+              ***** please fill all the fields *****
+            </p>
+          )}
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleClose}>
+            CANCEL
+          </Button>
+          <Button variant="primary" onClick={alerterror}>
+            CONFIRM
+          </Button>
+        </Modal.Footer>
+      </Modal>
+    </>
   );
 };
